+++ data/requestCounts.json
@@ -1,1 +1,1 @@
-{"2024-07-02":306,"TotalRequestCount":306}
+{"2024-07-02":306,"TotalRequestCount":325,"2024-07-03":19}
